requirements.txt
Django==1.10.5
djangorestframework==3.6.4
pkg-resources==0.0.0
